api/openai/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pypdf>=3.17.0
